reorganise files and folders, add SESSION_SECRET in env
