+++ pong_backend/src/app.module.ts
@@ -7,13 +7,12 @@
 import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { HelloController } from './hello/hello.controller';
-import { DatabaseModule } from './database.module';
+import { DatabaseModule } from './modules/database/database.module';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { User } from './models/orm_models/user.entity';
 import { Blocked } from './models/orm_models/blocked.entity';
 import { Friend } from './models/orm_models/friend.entity';
-import { UsersModule } from './users/users.module';
+import { UsersModule } from './modules/user/users.module';
 import * as session from 'express-session';
 import { Match } from './models/orm_models/match.entity';
 import { MatchHistory } from './models/orm_models/matchHistory.entity';
@@ -38,7 +37,7 @@ import { Channel } from './models/orm_models/channel.entity';
       Channel,
     ]),
   ],
-  controllers: [AppController, HelloController],
+  controllers: [AppController],
   providers: [AppService],
 })
 export class AppModule {}
