--- pong_backend/src/controllers/user/user.controller.ts
@@ -1,30 +1,0 @@
-import { Controller, Get, Post, Delete, Body, Param } from '@nestjs/common';
-import { ApiTags } from '@nestjs/swagger';
-import { User } from 'src/models/orm_models/user.entity';
-import { UserService } from 'src/services/userservice';
-
-@ApiTags('User')
-@Controller('user')
-export class UsersController {
-  constructor(private readonly userService: UserService) {}
-
-  @Get()
-  async findAll(): Promise<User[]> {
-    return this.userService.findAll();
-  }
-
-  @Get(':id')
-  async findOne(@Param('id') id: number): Promise<User> {
-    return this.userService.findOne(id);
-  }
-
-  @Post()
-  async create(@Body() user: User): Promise<User> {
-    return this.userService.create(user);
-  }
-
-  @Delete(':id')
-  async remove(@Param('id') id: number): Promise<void> {
-    await this.userService.remove(id);
-  }
-}
