--- pong_backend/src/hello/hello.controller.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { HelloController } from './hello.controller';
-
-describe('HelloController', () => {
-  let controller: HelloController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [HelloController],
-    }).compile();
-
-    controller = module.get<HelloController>(HelloController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
