+++ pong_backend/src/modules/profile/interfaces/friend_profile.interface.ts
@@ -1,0 +1,10 @@
+export class FriendProfile {
+  userId: number;
+  nickname: string;
+  avatar: string;
+  wins: number;
+  losses: number;
+  ladderLevel: number;
+  status: string;
+  achievements: string;
+}
