+++ pong_backend/src/modules/profile/interfaces/private_profile.interface.ts
@@ -1,0 +1,10 @@
+export class PrivateProfile {
+  userId: number;
+  nickname: string;
+  avatar: string;
+  wins: number;
+  losses: number;
+  ladderLevel: number;
+  status: string;
+  achievements: string;
+}
