+++ pong_backend/src/modules/profile/interfaces/public_profile.interface.ts
@@ -1,0 +1,7 @@
+export class PublicProfile {
+  userId: number;
+  nickname: string;
+  avatar: string;
+  ladderLevel: number;
+  achievements: string;
+}
