+++ pong_backend/src/modules/session/shared-session.service.ts
