+++ pong_backend/src/modules/user/users.module.ts
@@ -3,10 +3,10 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { User } from 'src/models/orm_models/user.entity';
 import { Blocked } from 'src/models/orm_models/blocked.entity';
 import { Friend } from 'src/models/orm_models/friend.entity';
-import { ProfileController } from './controllers/profile/profile.controller';
-import { ProfileService } from './services/profile/profile.service';
-import { UserController } from './controllers/user/user.controller';
-import { SharedSession } from 'src/shared/services/session/shared-session.service';
+import { ProfileController } from '../profile/profile.controller';
+import { ProfileService } from '../profile/profile.service';
+import { UserController } from './user.controller';
+import { SharedSession } from '../session/shared-session.service';
 
 @Module({
   imports: [TypeOrmModule.forFeature([User, Blocked, Friend])],
