--- pong_backend/src/shared/interfaces/user/friend_user.interface.ts
@@ -1,5 +1,0 @@
-export class FriendUser {
-    userId: number;
-    nickname: string;
-    status: string;
-  }
