--- pong_backend/src/shared/interfaces/user/public_user.interface.ts
@@ -1,4 +1,0 @@
-export class PublicUser {
-    userId: number;
-    nickname: string;
-  }
