--- pong_backend/src/users/interfaces/profile/friend_profile.interface.ts
@@ -1,10 +1,0 @@
-export class FriendProfile {
-    userId: number;
-    nickname: string;
-    avatar: string;
-    wins: number;
-    losses: number;
-    ladderLevel: number;
-    status: string;
-    achievements: string;
-  }
