--- pong_backend/src/users/interfaces/profile/public_profile.interface.ts
@@ -1,7 +1,0 @@
-export class PublicProfile {
-    userId: number;
-    nickname: string;
-    avatar: string;
-    ladderLevel: number;
-    achievements: string;
-  }
